chore: reorganized folder arhitecture and updated parsing pipeline
