+++ ingestion/html_scraper.py
@@ -1,6 +1,10 @@
+from __future__ import annotations
+
+from typing import Optional
+
 import requests
 from bs4 import BeautifulSoup
-from typing import Optional
+
 
 DEFAULT_HEADERS = {
     "User-Agent": (
@@ -13,33 +17,21 @@ DEFAULT_HEADERS = {
 
 def scrape_html_source(url: str) -> Optional[str]:
     """
-    Fetches the raw HTML source code from the given URL.
+    Fetch raw HTML source from the given URL.
 
-    Args:
-        url (str): The link to scrape.
-
-    Returns:
-        Optional[str]: The raw HTML string if successful, None otherwise.
+    This helper only fetches HTML. Canonical parsing happens in
+    scripts/run_parser_pipeline.py via html_cleaner + StructuralParser.
     """
     try:
         response = requests.get(url, headers=DEFAULT_HEADERS, timeout=30)
         response.raise_for_status()
         return response.text
-    except requests.exceptions.RequestException as e:
-        print(f"[scraper] Error fetching URL {url}: {e}")
+    except requests.exceptions.RequestException as exc:
+        print(f"[scraper] Error fetching URL {url}: {exc}")
         return None
 
 
 def scrape_with_soup(url: str) -> Optional[BeautifulSoup]:
-    """
-    Fetches a webpage and parses it directly with BeautifulSoup.
-
-    Args:
-        url (str): The link to scrape.
-
-    Returns:
-        Optional[BeautifulSoup]: A parsed BeautifulSoup object, or None on failure.
-    """
     html = scrape_html_source(url)
     if html is None:
         return None
