+++ scripts/export_canonical_bundle.py
@@ -13,8 +13,10 @@ from ingestion.exporters import export_canonical_bundle
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(description="Export a canonical parser bundle from legacy units")
-    parser.add_argument("--input", required=True, help="Path to legacy legal_units.json")
+    parser = argparse.ArgumentParser(
+        description="Export a canonical parser bundle from structural intermediate units"
+    )
+    parser.add_argument("--input", required=True, help="Path to intermediate legal_units.json")
     parser.add_argument("--out-dir", required=True, help="Output directory for canonical bundle")
     parser.add_argument("--law-id", required=True, help="Canonical law id, e.g. ro.codul_muncii")
     parser.add_argument("--law-title", required=True, help="Law title, e.g. Codul muncii")
