+++ tests/ingestion/test_html_scraper.py
@@ -1,6 +1,6 @@
 import pytest
 from unittest.mock import patch, MagicMock
-from ingestion.parser.html_scraper import scrape_html_source, scrape_with_soup
+from ingestion.html_scraper import scrape_html_source, scrape_with_soup
 from bs4 import BeautifulSoup
 import requests
 
@@ -9,7 +9,7 @@ import requests
 
 def test_scrape_html_source_success():
     mock_html = "<html><body><h1>Test</h1></body></html>"
-    with patch("ingestion.parser.html_scraper.requests.get") as mock_get:
+    with patch("ingestion.html_scraper.requests.get") as mock_get:
         mock_response = MagicMock()
         mock_response.text = mock_html
         mock_response.raise_for_status.return_value = None
@@ -21,7 +21,7 @@ def test_scrape_html_source_success():
 
 
 def test_scrape_html_source_failure():
-    with patch("ingestion.parser.html_scraper.requests.get") as mock_get:
+    with patch("ingestion.html_scraper.requests.get") as mock_get:
         mock_get.side_effect = requests.exceptions.RequestException("Connection error")
 
         result = scrape_html_source("https://example.com")
@@ -33,7 +33,7 @@ def test_scrape_html_source_failure():
 
 def test_scrape_with_soup_returns_beautifulsoup():
     mock_html = "<html><body><p>Hello</p></body></html>"
-    with patch("ingestion.parser.html_scraper.requests.get") as mock_get:
+    with patch("ingestion.html_scraper.requests.get") as mock_get:
         mock_response = MagicMock()
         mock_response.text = mock_html
         mock_response.raise_for_status.return_value = None
@@ -46,7 +46,7 @@ def test_scrape_with_soup_returns_beautifulsoup():
 
 
 def test_scrape_with_soup_returns_none_on_failure():
-    with patch("ingestion.parser.html_scraper.requests.get") as mock_get:
+    with patch("ingestion.html_scraper.requests.get") as mock_get:
         mock_get.side_effect = requests.exceptions.ConnectionError()
 
         result = scrape_with_soup("https://example.com")
